Restrict network key file permissions
