+++ util/src/keys/directory.rs
@@ -18,6 +18,7 @@
 
 use common::*;
 use std::path::{PathBuf};
+use path::restrict_permissions_owner;
 
 const CURRENT_DECLARED_VERSION: u64 = 3;
 const MAX_KEY_FILE_LEN: u64 = 1024 * 80;
@@ -463,23 +464,6 @@ pub struct KeyDirectory {
 	path: String,
 	cache: RwLock<HashMap<Uuid, KeyFileContent>>,
 	cache_usage: RwLock<VecDeque<Uuid>>,
-}
-
-/// Restricts the permissions of given path only to the owner.
-#[cfg(not(windows))]
-pub fn restrict_permissions_owner(file_path: &Path) -> Result<(), i32>  {
-	let cstr = ::std::ffi::CString::new(file_path.to_str().unwrap()).unwrap();
-	match unsafe { ::libc::chmod(cstr.as_ptr(), ::libc::S_IWUSR | ::libc::S_IRUSR) } {
-		0 => Ok(()),
-		x => Err(x),
-	}
-}
-
-/// Restricts the permissions of given path only to the owner.
-#[cfg(windows)]
-pub fn restrict_permissions_owner(_file_path: &Path) -> Result<(), i32>  {
-	//TODO: implement me
-	Ok(())
 }
 
 impl KeyDirectory {
